cli: Refactored out Vagrant SSH detection to driver
